requirements.txt
langchain
langchain_community
langchain-google-genai
chromadb
gradio
pypdf
python-dotenv
